requirements.txt
gradio==4.26.0
litellm==1.35.12
python-dotenv==1.0.1
youtube_transcript_api==0.6.1
